Former-commit-id: 1fd7d980151bb22031a95fd9e6f0f8d96eba46de

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/assessment/transactional/AssessmentServiceImpl.java
@@ -107,7 +107,7 @@ public class AssessmentServiceImpl extends BaseTransactionalService implements A
     }
 
     @Override
-    public ServiceResult<Long> countByStateAndCompetition(AssessmentStates state, long competitionId) {
+    public ServiceResult<Integer> countByStateAndCompetition(AssessmentStates state, long competitionId) {
         return serviceSuccess(assessmentRepository.countByActivityStateStateAndTargetCompetitionId(state.getBackingState(), competitionId));
     }
